Remove json_t usage from GetExportedFunctions
As well as the mod_func functions.

Due to the way I decided to do this, thcrap now requires C++17. This is
fine in my opinion because there are already parts in the thcrap
codebase using std::filesystem::path which is a C++17 feature. I needed
C++17 for std::string_view
# Bitte geben Sie eine Commit-Beschreibung für Ihre Änderungen ein. Zeilen,
# die mit '#' beginnen, werden ignoriert, und eine leere Beschreibung
# bricht den Commit ab.
#
# Auf Branch runconfig_struct
# Ihr Branch ist auf demselben Stand wie 'origin/runconfig_struct'.
#
# Zum Commit vorgemerkte Änderungen:
#	geändert:       Base.props
#	geändert:       thcrap/src/init.cpp
#	geändert:       thcrap/src/pe.cpp
#	geändert:       thcrap/src/pe.h
#	geändert:       thcrap/src/plugin.cpp
#	geändert:       thcrap/src/plugin.h
#	geändert:       thcrap/src/runconfig.cpp
#
# Änderungen, die nicht zum Commit vorgemerkt sind:
#	geändert:       thcrap/src/init.cpp
#