netcode: improve logs
- We don't want to print a message about cancelled downloads,
they usually means the file was successfully downloaded from
another server.
- If the same file is downloaded from 2 servers in parallel,
the download count was updated for every finished download,
so it becomes too high. The same error probably happens with
failed downloads. We now change the downloaded files count
only on successful downloads.
- The wording for the files.js download error was rather bad,
the "patch_name files.js" part was confusing. And we can get
the patch name from the URL, so ommiting it in the message
seems better.