thcrap_update: lock containers access behind a lock
futuresList had a theorical race condition where 2 threads could
call push_back at the same time, messing the vector internal state.
And files had an actual race condition that could be triggered
reliably during testing by just downloading many files.
If several files.js downloads end roughly at the same time, they
would be calling addFile at the same time.
The addFile function wants to do 3 things: create a new file,
add it to vector, and add it to the queue (which means, add a task
to the thread pool and add the returned future to the futuresList).
We did that by first `emplace_back`ing the file into the vector,
retrieving the element we just added with files.back(), and we
add that file to the queue.
Between the emplace_back call and the back() call, another thread
could also call emplace_back. The 2 threads would then get the
same file from the back() call, and try to add it 2 times to the
queue. The file detects that error and throws an exception.