netcode: improve error reporting.
Before, we were reporting errors at the lowest layer. This layer
knows how the download failed, so it can print the more informations
about it. But with this architecture, the callers con't have a lot
of control over when / how the errors are displayed. Also, the
lower layer don't have all the context, it only has the URL.
And, handling errors in the lower layer makes the error handling
in the upper layer a bit more awkward. Either they dispay their own
error messages, which mean the error is displayed twice, or they
display messages only for their own errors, which means "I print
an error for this case but not for that one because someone
already printed it".

Now, the errors are always displayed at a high layer (either the
download callback, update.cpp or repo_discovery.cpp), and the lower
layer return a precise error message to the upper layers through
a return value or the error callback.