thcrap_update: don't download the same file from 2 sources in parallel
This seemed like a good idea at first - if a server is dead or
really slow, we go to the 2nd one directly, without waiting for
the timeout or the end of the really slow download.
But in normal cases (when all the servers are alive and with a
nice download speed), we download the file 2 times, which means
our download speed is halved. This also put an extra load on the
servers.

And removing this feature means that we no longer need to synchronize
multiple downloads of the same file, which leads to the removal of
almost all the synchronization code from the File class.
We can also remove the tight coupling between a File and its
Downloader. When a File doesn't need someone to schedule its
multiple downloads, it can just try every server itself.