thcrap: mitigate a race condition in BP_fragmented_file_read
Fixes a crash in AoCF. What happened was that the game was trying
to read the same file from 2 threads simultaneously. When the 1st
thread enters BP_fragmented_file_read, it allocates a rep_buffer
and does things with it. At the same time, another thread would
enter BP_fragmented_file_read with the same file_rep, notice that
a rep_buffer is available, and start using it.
Then, the 1st thread finishes patching the file, frees its file_rep
and set it to NULL. And then, the 2nd thread crashes because someone
just freed its file_rep.

For now, I fixed it by just adding a mutex. A real fix would be to
make every thread have its own file_rep. Ideally, we would remove
the file_rep cache built in BP_file_header and build them on-demand,
only for the thread using them. This would speed up launching time
(and in debug mode, it would be *a lot* faster because printing all
these "we tried to load this file" takes a lot of time), and, if
done right, it would let modders add new files instead of only being
able to replace existing ones.

I also enabled process termination on heap corruption, because a
crash at the point of memory corruption is easier to debug than
random things happening because of memory corruption.