thcrap_tasofro: defer filenames writing to a background thread
The filenames usually arrive in bursts, so dumping the fileslist.js
for every new filename is quite wasteful. With a consumer thread,
we can wait for the producer to produce a few filenames and then
write them when the burst is finished.

This is used only when dat_dump is enabled.

Also fixes a bug where the current directory changes between 2
register_filename calls (this bug was actually the reason I started
working on this feature).