Fix some warnings on MinGW
- Some things about unused variables and parameters.

- Replacing __declspec(align(16)) with the standard alignas(16).
Even the msdn page on __declspec(align()) recommends to do this.

- Change the thcrap_update_wrapper type assertion.
The old way I was doing this is by declaring a variable with the
type of the original pointer, and assign it a pointer to the new
function. I replaced it with a static_assert that compares the
types of the 2 pointers, which has some advantages:
  * It is actually designed for that, instead of exploiting an
    error during compilation.
  * It doesn't declare a new variable that ends up being unused.
  * It works, while with the old way, the unused variable was
    discarded by g++ before it even noticed the type mismatch
	(and this kind of optimization could also happen with MSVC in
	a future version).

- Fix a bug in jsonvfs_add where any backslash in the input string
would lead to an infinite loop.