thcrap: remove patch.js
In the future, we will want to display more informations in
thcrap_configure during the patch selection.
When we download the initial patch list, we download only the repo.js
for every repository. We could download the patch.js for every
patch in the thcrap network, but that would mean download a lot of
files (more than 100), potentially from a lot of different servers,
only to display a list of patch. That sounds needlessly slow.
Instead, we decided to include everything we care about in repo.js.
And, because having these fields duplicated in 2 files is harder to
maintain, we remove patch.js, and thcrap will manage to get
everything it needs from repo.js or other files instead.

Reading from patch.js is still supported as a fallback for older patches.