thcrap_tasofro: rework file patching
The old implementation had this huge function that did everything
and that became impossible to maintain.

This new implementation tries to put more of the game-specific code
in game-specific files, and to add more breakpoints instead of
putting as many things as possible in a single do-everything
breakpoint.

I also moved it from the thcrap core dll to the tasofro plugin,
because having a C++ interface and being able to pass lambdas
as parameters really helps.

And finally, I changed some breakpoints to patch files when the game
loads them, instead of patching all the game's files at startup.
This should decrease RAM usage and loading time at startup, but
increases loading time when the game loads files.
IIRC, I made this change for the nsml and megamari games. th105 and
th123 were already using this, and the th135 games are still loading
everything at startup.