build: remove timestamping when signing.
Timestamping relies on an external server to provide a timestamp
dll. This means building thcrap requires an internet connection.
Also, the URL we were using for the timestamp dll died (it was
actually deprecated for a few years).

And we don't really need timestamping. Windows doesn't care about
our certificate because it's self-signed, and we only care about
checking the signature of new releases.