thcrap_wrapper: use LoadStringW without reallocing
If you pass 0 as buffer size, LoadStringW returns the actual string
size. It also returns a pointer to the (unterminated) string in .rsrc
section.

This feature was documented at some point between January 2007 and
March 2017, but existed long before that.

Windows NT: absent in Win NT 4.0 RTM, present in Win 2000 RTM
Wine: added in 0.9.58 (March 2008) [1]
Windows 9x: no (UnicoWS)

By the way, this is specific to the Unicode version of LoadString.

[1]: <https://source.winehq.org/git/wine.git/commitdiff/9978cc9c56f7ec09be2d2a22c2e7fb5d5d1f07be>