compile with /utf-8
A moment of silence for "warning C4819: The file contains a character
that cannot be represented in the current code page (932). Save the file
in Unicode format to prevent data loss"

This flag only affects how the compiler reads the source code (source
charset) and how it produces the bytes for narrow strings (execution
charset). Before VS2015 it would always convert from the source charset
(detected using BOM) to CP_ACP. Now we can make it read and emit UTF-8,
and we don't even need to use BOM.