thcrap_tasofro: fix file patching for th135
The old function need_replace was first used to check whether we need to read
the original file from the game's archive. Then, I noticed it was similar to
the one-line test we use in a lot of places to check if we want to patch a
file, and I decided to reuse that function there.

But this function isn't identical to the one-line check, there is a small
difference. And because of that, file replacing stopped working in
th135-style games.

I moved the old implementation to a new, private function (need_orig_file)
which returns true if we need to read the original file, and in need_replace,
I put the correct implementation.