Small improvements to wrapper utils.
Fixed the "unresolved external symbol _memset" error with the v142 compiler by
restructuring the relevant loop in my_memset so that the compiler doesn't think
it could replace it with a normal memset call.

Fixed that the return value of my_wcscmp was opposite that of the standard
wcscmp, which was previously unnoticed since all uses merely compared != 0.

All changes to function structure were tested to generate more efficient
assembly and verified to produce identical results, even in edge cases.