thcrap: Minor housekeeping/cleanup/optimizations
Apparently strdup is allowed to segfault when passed a null pointer. The
windows implementation behaves nicely so it probably isn't an issue, but
it seems like a good idea to not rely on it.
Also slightly optimized a few commonly used string literal copies so the
compiler can more intelligently convert them to known-size MOV operations.