thcrap_tasofro: rework th155 file patching
The old code was doing a lot of work when the game parses the archive header
(so, when the game starts), and not much after that. This meant the patched
game was way slower to start than the original (need to resolve *every* file
at startup), used a lot more RAM than the original (we load everything at
startup and keep them for whenever we need it), and didn't support repatching.

The new code patch files when the game actually opens them, and free the
file_rep when the game closes them. This means that starting a fight is
slightly slower (because we have to patch the files there), but it fixes all
the issues above.