thcrap_tasofro: fix encoding-related crash on fileslist loading
We were giving a sjift-jis string to std::filesystem::path::operator=,
but it takes a string in the ANSI code page (and tries to convert that
string to UTF-16).
It works when the ANSI codepage is latin-1 or another single-byte encoding,
but MBCS have a bunch of invalid characters that can be used to detect
invalid strings, and fs::path decides to throw an exception when it sees
them.