thcrap: Fix recursive error log failure on Wine
For some reason wine isn't locating the message table resource in ntdll, thus causing the exception handler to dereference a NULL pointer and repeatedly invoke itself until the stack overflows. This "fix" merely checks for when that happens to avoid the crash, so more work will need to be done to make descriptions work properly on wine.