thcrap: experiment - WinAPI async logging
The point of this experiment is to try and address an issue where due to
a bug in condition variables themselves, thcrap would hang on shutdown
on Windows 7 due to the log thread not terminating.

https://stackoverflow.com/questions/60452341/stdconditional-variable-notify-problem-on-windows-7-called-from-destructor-o

In the condition variable, the issue seems to boil down to
ZwReleaseKeyedEvent. I use a semaphore in this code so it shouldn't be
an issue.