Build: disable WholeProgramOptimization for wrappers
For some reason, in release mode, the linker would fail to link the
executable, complaining about missing a _memset symbol in
install_popup.obj. When I tried to disassemble the file, it didn't have
any code in it. It took me months to figure out that whole program
optimization moves all code generation to the linking stage and that the
.obj files can not be disassembled when whole program optimization is
enabled. So I disabled whole program optimization to get .obj files that
I can actually disassemble but that fixed the problem. I have decided to
push this because the wrappers don't need whole program optimization
anyways