configure: add support for using wrapper instead of shortcuts
No UI yet, I'm not sure how we want to present the setting for that, and it's
not completely ready (it still has one bug), but at least the implementation
is there for whenever we decide to do something with it.
This can be used by using the configure_v3 and by changing
default_shortcut_destination. That value is a bitmap but in decimal because
json, have fun doing conversions with it! Using shortcuts (default) is 65536,
using a wrapper with absolute path is 65536*2=131072, and using a wrapper with
relative path is 65536*3=196608. Note that the configure_v3 will also add a few
bits at the end of this number.