thcrap_tasofro: fix encoding error in bmpfont generation
When generating bmpfont, we would use json_load_file (which then uses fopen)
with an UTF-8 path to load json file while choosing which characters we put
in the bmp font. If this path contains non-ASCII characters (which happens
if the path to thcrap contains non-ASCII characters), we would silently fail
to load every json file, which results in the bmpfont containing only the
characters from the original in-game bmpfont.

We now read the file with file_read instead, which supports UTF-8 paths
properly.