configure: declare game_result_t.descriptor as IntPtr instead of string
It's an UTF-8 string, but C# string type will marshal it as an ANSI string.
I tried with UnmanagedType.LPUTF8Str and .NET Framework 4.8, which seems to
be designed exactly for this, but for some reason, it gives me garbage.
And actually, we never use this field, so we can just declare it as IntPtr
so that it doesn't get marshalled in anyway, and ignore it.