update: download th105 files for th123
The th123 game can be set up to depend on th105 files, and when it is, thcrap
needs the th105 patch files.

It would be better to have something generic instead of hardcoding that one
dependency, but we only have one dependency like that in all our supported
games, and making something generic is more work.
If we ever need a dependency like that for another game, we'll revert this
commit and implement something nicer.

Fixes #203