thcrap_tasofro: add bidi support for arabic
The Tasofro games don't use Windows' text rendering engine and implement their
own, which of course don't handle arabic text correctly.

In these games, we have 2 challenges:
- displaying the text right to left;
- converting from general form to contextual form (https://en.wikipedia.org/wiki/Arabic_script_in_Unicode).
The FriBidi library handle both problems, so this is mostly about adding that
lib, using it in every place we need it, and helping the lib with interpreting
some tags properly.