search: use publisher name instead of DisplayName
The DisplayName can change between regions. For example, ZUN's games will use
"東方" when the system language is anything but English, where it will use
"Touhou" instead. It also isn't really reliable, with games like Touhou 14.3
and Touhou 18.5 not starting with "東方" (we had to make an exception for
Touhou 14.3, and Touhou 18.5 actually wasn't supported because of this).

The Publisher field doesn't seem to change between languages, and it will
definitely not change if one game decides to have a different name.