scripts: parse json with json5 in repo_update.py
Quite often, people are testing their patches locally with thcrap, using
JSON files with trailing commas, and it works because thcrap uses a JSON5
library for JSON parsing, and then I try to put them on our mirror server
and we spend like 3 rounds of me telling them to fix something and them
fixing it before their file is finally compatible with the regular
JSON parser that repo_update.py uses.
Now, the server will support JSON5 as well, so their JSON file will be
compatible with the server.
We keep using the regular json library if the json5 one isn't available
in order to keep things working the same for people running the script
on their computers.