thcrap: try to automatically create a portable installation when possible
There are 2 major things that prevents thcrap from being portable: the
shortcuts have absolute paths, and we have paths pointing towards the games.

For paths pointing towards the games, we'll make them relative instead of
absolute if the games are close enough to thcrap and both are likely to be
moveable together.

For shortcuts, we'll replace them with small executables that run thcrap_loader
with a relative path and the correct arguments when they're either in thcrap's
directory or in the games' directory. We'll also always use them on wine,
because wine tends to have some problems with the shortcuts (especially if you
want to configure your Steam Deck to run the games from your Steam library).