thcrap_tasofro: fill the overallocated part of rep_buffer with zeroes
When we replace a file with a smaller one, we overallocate our rep buffer
because we always give the the game the biggest of the 2 sizes.
But we used to only fill the useful part of the rep buffer with the rep
file, and keep the overallocated part as garbage, which then ends up in the
game's internal buffer.
This is fine for a bunch of file types that have clear end markers, like png
where the header contains the file size or compiled nut scripts which end with
a little endian TAIL tag, but it broke things when replacing plaintext nut
scripts because the nut compiler tried to interpred the garbage at the end as
code.