configure_v3: fix race condition in auto search
If the auto search finds something, we want to make the next button clickable.
If it doesn't find anything, we want it to stay greyed out.

Because I'm kind of inexperimented with this async stuff, the code that checks
whether any game was found and deciding on whether the "next" button should be
clickable or grayed out could be running *before* the actual search was
completed, and then of course the result was "I didn't find any game (yet)".