configure_v3: rename DLL on save and deal with both DLLs if present
Like the title says, rename behavior has been changed,
now only renaming the DLL when the user saves, behaving like the other options.
I also implemented a fix for if both the enabled and disabled DLLs are present,
now deleting the disabled one and defaulting to enabled in such a case.