thcrap_tasofro: pl: fix child versions of parseCommand not being called
Commit 068829b changed a parameter for AText::parseCommand from an int to a
size_t. But it didn't apply the same changes to StoryText::parseCommand and
Th155StoryText::parseCommand, which were overriding that function.
Because of that, the 2 functions from the child classes were no longer
overriding the one from the parent class.

Fixes an issue in th145 where only the last line of dialog balloons was
displayed, and may fix #286