thcrap_tasofro: automatically fix th155 win files
Some files miss the column for the 2nd English balloon, and all of them miss
the column for the 3rd English balloon.
This is fine in the base game because the files without the 2nd English balloon
only ever use 1 balloon for every dialogue, but of course thcrap breaks this
assumption.